Revamp landing page
- Shift to single-page structure, removing sections
- Transition to inline styling for fidelity
- Add language toggle feature
- Consolidate assets into refined hierarchy
