+++ landing/index.html
@@ -8,16 +8,19 @@
     <link rel="icon" type="image/x-icon" href="/favicon.ico" />
     <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
     <link rel="manifest" href="/site.webmanifest" />
-    <meta name="theme-color" content="#3b82f6" />
-    <title>EndurAI — AI-powered triathlon coach</title>
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700;800&display=swap" rel="stylesheet" />
+    <meta name="theme-color" content="#faf9f5" />
+    <title>Endurai — Intelligent endurance analytics</title>
     <meta
       name="description"
-      content="Recovery score, HRV analysis and AI recommendations every morning in Telegram"
+      content="Recovery score, HRV analysis and a daily plan — every morning in Telegram. Open source with an MCP server. Runs on your Intervals.icu data."
     />
-    <meta property="og:title" content="EndurAI — AI-powered triathlon coach" />
+    <meta property="og:title" content="Endurai — Intelligent endurance analytics" />
     <meta
       property="og:description"
-      content="Recovery score, HRV analysis and AI recommendations every morning in Telegram"
+      content="Recovery score, HRV and a daily plan — every morning in Telegram. Open source + MCP server."
     />
     <meta property="og:image" content="https://endurai.me/og-image.png" />
     <meta property="og:url" content="https://endurai.me" />
