+++ landing/package.json
@@ -16,9 +16,6 @@
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react": "^4.3.3",
-    "autoprefixer": "^10.4.20",
-    "postcss": "^8.5.14",
-    "tailwindcss": "^3.4.15",
     "typescript": "~5.6.3",
     "vite": "^6.0.1"
   }
