+++ landing/src/index.css
@@ -1,41 +1,18 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-:root {
-  --bg: #ffffff;
-  --surface: #f5f5f7;
-  --surface-2: #ebebf0;
-  --border: #d4d4dc;
-  --text: #1a1a2e;
-  --text-dim: #6b6b80;
-  --accent: #2563eb;
-  --accent-glow: #2563eb18;
-  --green: #166534;
-  --yellow: #92400e;
-  --orange: #9a3412;
-  --red: #991b1b;
-}
-
 * {
   box-sizing: border-box;
   margin: 0;
   padding: 0;
 }
 
+html,
+body {
+  height: 100%;
+}
+
 body {
   font-family: 'Inter', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  background: var(--bg);
-  color: var(--text);
+  background: #faf9f5;
+  color: #0a0d18;
   line-height: 1.5;
   -webkit-font-smoothing: antialiased;
-}
-
-@keyframes fadeIn {
-  from { opacity: 0; }
-  to { opacity: 1; }
-}
-
-.animate-fadeIn {
-  animation: fadeIn 0.15s ease-out;
 }
