--- landing/tailwind.config.js
@@ -1,25 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-  content: ['./index.html', './src/**/*.{ts,tsx}'],
-  theme: {
-    extend: {
-      colors: {
-        bg: 'var(--bg)',
-        surface: 'var(--surface)',
-        'surface-2': 'var(--surface-2)',
-        border: 'var(--border)',
-        text: 'var(--text)',
-        'text-dim': 'var(--text-dim)',
-        accent: 'var(--accent)',
-        green: 'var(--green)',
-        yellow: 'var(--yellow)',
-        orange: 'var(--orange)',
-        red: 'var(--red)',
-      },
-      fontFamily: {
-        sans: ['Inter', '-apple-system', 'BlinkMacSystemFont', 'Segoe UI', 'sans-serif'],
-      },
-    },
-  },
-  plugins: [],
-}
